requirements.txt
numpy
xlrd
PyQt5
matplotlib
ruamel-yaml
pandas
tqdm